added object's owners and tokens

+++ src/main/java/org/example/labbb1/services/ChapterService.java
@@ -1,6 +1,8 @@
 package org.example.labbb1.services;
 
+import org.example.labbb1.exceptions.ForbiddenException;
 import org.example.labbb1.model.Chapter;
+import org.example.labbb1.model.User;
 import org.example.labbb1.repositories.ChapterRepository;
 import org.example.labbb1.repositories.CoordinatesRepository;
 import org.example.labbb1.repositories.SpaceRepository;
@@ -34,8 +36,15 @@ public class ChapterService {
     public void addNewChapter(Chapter chapter) throws PSQLException {
         chapterRepository.save(chapter);
     }
-    public void updateChapter(Chapter chapter) throws PSQLException{
-        chapterRepository.save(chapter);
+    public boolean updateChapter(Chapter chapter, User user) throws PSQLException, ForbiddenException{
+        var chap = chapterRepository.findById(chapter.getId());
+        if(chap.isPresent()) {
+            Chapter chapter1 = chap.get();
+            if(chapter1.getUser().getId().equals(user.getId())) {
+                chapterRepository.save(chapter);
+                return true;
+            } else throw new ForbiddenException();
+        } else return false;
     }
 
     public Iterable<Chapter> getAllChapters(){
@@ -60,10 +69,15 @@ public class ChapterService {
     }
 
 
-    public void deleteChapter(Long id){
-//        Chapter chapter = new Chapter();
-//        chapter.setId(id);
-        chapterRepository.deleteById(id);
+    public boolean deleteChapter(Long id, User user) throws ForbiddenException{
+        var chap = chapterRepository.findById(id);
+        if(chap.isPresent()) {
+            Chapter chapter = chap.get();
+            if(chapter.getUser().getId().equals(user.getId())) {
+                chapterRepository.deleteById(id);
+                return true;
+            } else throw new ForbiddenException();
+        } else return false;
     }
 
 }
